Added recommended doctors display
