+++ eBolnica/eBolnica.Model/Models/Korisnik.cs
@@ -12,8 +12,8 @@ namespace eBolnica.Model.Models
         public string Email { get; set; } = null!;
 
         public string KorisnickoIme { get; set; } = null!;
-          public byte[]? Slika { get; set; }
-         public byte[]? SlikaThumb { get; set; }
+        public byte[]? Slika { get; set; }
+        public byte[]? SlikaThumb { get; set; }
         public DateTime? DatumRodjenja { get; set; }
         public string? Telefon { get; set; }
         public string? Spol { get; set; }
